requirements.txt
Django==5.0.6
djangorestframework==3.15.2
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
django-filter==24.2
reportlab==4.1.0
psycopg[binary]==3.2.11
httpx==0.27.0
django-cors-headers==4.4.0
